added open graph for sharing profile URL
